+++ src/backbone.marionette.collectionview.js
@@ -121,9 +121,9 @@ Marionette.CollectionView = Marionette.View.extend({
 
   // Build an `itemView` for every model in the collection. 
   buildItemView: function(item, ItemView){
-    var view = new ItemView({
-      model: item
-    });
+    var itemViewOptions = getAttribute(this, "itemViewOptions");
+    var options = _.extend({model: item}, itemViewOptions);
+    var view = new ItemView(options);
     return view;
   },